third_party/blink/web_tests/http/tests/security/isolatedWorld/bypass-main-world-csp-iframes-expected.txt

ALERT: Running test #6

CONSOLE ERROR: line 28: Refused to frame 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=6' because it violates the following Content Security Policy directive: "frame-src 'none'".

ALERT: Running test #5

CONSOLE ERROR: Refused to frame 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=5' because it violates the following Content Security Policy directive: "frame-src 'none'".

ALERT: Running test #4

ALERT: Starting to bypass main world's CSP:
ALERT: iframe loaded in test #4
ALERT: Running test #3

CONSOLE ERROR: line 28: Refused to frame 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=3' because it violates the following Content Security Policy directive: "frame-src 'none'".

ALERT: iframe loaded in test #3.5
ALERT: Running test #2

CONSOLE ERROR: line 28: Refused to frame 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=2.5' because it violates the following Content Security Policy directive: "frame-src 'none'".

ALERT: Running test #1

CONSOLE ERROR: line 28: Refused to frame 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=1' because it violates the following Content Security Policy directive: "frame-src 'none'".

ALERT: Running test #0

 This test ensures that iframes loaded from isolated worlds marked with their own Content Security Policy aren't affected by the page's content security policy. Extensions, for example, should be able to load any resource they like.
